# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: VantaDB HTTP API
  version: "0.6.0"
  description: |
    REST interface for the VantaDB HTTP server.

    VantaDB is a high-performance vector-native knowledge graph database
    with IQL (Interactive Query Language) for semantic memory operations,
    hybrid search, and graph traversal.

    The HTTP server is optional and enabled via `vanta-cli server --http`
    or the `server` Cargo feature.

    Parity contract: every route registered in `src/server/router.rs`
    (`app_with_cors` + dashboard fallback via `mount_dashboard`) must appear
    here with matching methods — enforced by `scripts/check_openapi_parity.mjs`
    in CI (gate-docs-21). Endpoints marked `x-experimental: true` are unstable
    and may change without notice.
  contact:
    name: VantaDB Project
  license:
    name: Apache-2.0
servers:
  - url: http://127.0.0.1:8080
    description: Local development server (default)
  - url: https://{host}:{port}
    description: Production server with TLS
    variables:
      host:
        default: localhost
      port:
        default: "443"
...
paths:
...
  /dashboard:
    get:
      summary: Web console entry point
      description: |
        Serves the Vanta Studio console when started with `--dashboard-dir`;
        otherwise returns 404 with a hint to enable it (WEB-03).
      operationId: dashboardDisabled
      tags:
        - System
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard index.html (when --dashboard-dir configured)
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Dashboard not enabled
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /dashboard/{path}:
    get:
      summary: Web console static assets
      description: Static file serving fallback for the dashboard SPA.
      operationId: dashboardStaticPath
      tags:
        - System
      security:
        - bearerAuth: []
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Static asset or index.html fallback
        "404":
          description: Asset not found or dashboard not enabled
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v2/query:
    post:
      summary: Execute IQL query
      description: |
        Execute an IQL (Interactive Query Language) statement against the database.
        IQL keywords are case-sensitive UPPERCASE (`FROM`, `WHERE`, `SIGUE`,
        `INSERT`, `TYPE`, …) — lowercase variants fail with a parse error.
        Verified against `src/parser/mod.rs`.

        Supported operations are passed as a single IQL query string:

        ## Read Operations

        **Query (search):** Scan an entity with optional filters, vector similarity search, graph traversal, sorting, and projection.
        ```
        FROM <entity> [SIGUE <min>..<max> "<edge_label>" [TYPE <type>] [AS <alias>] [<alias>]]
          [WHERE <field> <op> <value> [AND <field> <op> <value> ...]]
          [FETCH <field>,...] [RANK BY <field> [DESC]]
          [WITH TEMPERATURE <float>] [ROLE "<role>"]
          [PROFILE keyword|vector|hybrid [rrf_k <n>] [candidate_k <n>]]
        ```
        (`MATCH <entity>` is accepted as an alias of `FROM <entity>`.)

        **Vector/hybrid search:** Include a vector similarity condition in the WHERE clause.
        ```
        FROM <entity> WHERE <field> ~ "<text_query>", min = <score>
        ```

        **Text search:** Match a text field without a score threshold.
        ```
        FROM <entity> WHERE <field> ~ "<text_query>"
        ```

        **Graph traversal:** Follow edges from matched nodes.
        ```
        FROM <entity> SIGUE <min>..<max> "<edge_label>" [TYPE <type>] [AS <alias>]
        ```

        ## Write Operations

        **Insert:** Create a new node.
        ```
        INSERT NODE#<id> TYPE <type> {<key>: <value>,...} [VECTOR [<f32>,...]]
        ```

        **Update:** Modify an existing node's fields or vector.
        ```
        UPDATE NODE#<id> SET <key> = <value>,...
        UPDATE NODE#<id> SET VECTOR [<f32>,...]
        ```

        **Delete:** Remove a node.
        ```
        DELETE NODE#<id>
        ```

        **Relate:** Create a weighted or unweighted edge between two nodes.
        ```
        RELATE NODE#<source_id> --"<label>"--> NODE#<target_id> [WEIGHT <f32>]
        ```

        **Insert Message (conversational sugar):** Create a threaded message node with automatic embedding.
        ```
        INSERT MESSAGE <SYSTEM|USER|ASSISTANT> "<content>" TO THREAD#<thread_id>
        ```
      operationId: executeQuery
      tags:
        - Query
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v2/records:
    post:
      summary: Put record
      description: Create or overwrite a memory record (upsert by namespace+key).
      operationId: putRecord
      tags:
        - Records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordInput"
      responses:
        "200":
          description: Record stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete records by metadata filter
      description: Deletes all records in a namespace whose metadata matches the JSON filter string passed as query parameter.
      operationId: deleteRecordsByFilter
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: query
          required: true
          schema:
            type: string
        - name: filter
          in: query
          required: true
          description: JSON-encoded metadata filter (`MemoryFilter`)
          schema:
            type: string
      responses:
        "200":
          description: Matching records deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/records/batch:
    post:
      summary: Put records batch
      description: Upsert multiple records in one call (array of `RecordInput`).
      operationId: putRecordsBatch
      tags:
        - Records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/RecordInput"
      responses:
        "200":
          description: Batch stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/records/{ns}/{key}:
    get:
      summary: Get record
      description: Fetch a single record by namespace and key.
      operationId: getRecord
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NamespacePath"
        - $ref: "#/components/parameters/KeyPath"
      responses:
        "200":
          description: The stored record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete record
      description: Soft-delete (tombstone) a record by namespace and key.
      operationId: deleteRecord
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NamespacePath"
        - $ref: "#/components/parameters/KeyPath"
      responses:
        "200":
          description: Record deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/records/{ns}/{key}/versions:
    get:
      summary: List record versions
      description: Version history of a record.
      operationId: listRecordVersions
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NamespacePath"
        - $ref: "#/components/parameters/KeyPath"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Version list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/list:
    get:
      summary: List records
      description: Paginated listing of records in a namespace, with optional metadata filter ops.
      operationId: listRecords
      tags:
        - Records
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: cursor
          in: query
          schema:
            type: string
        - name: filter_ops
          in: query
          description: JSON-encoded filter ops
          schema:
            type: string
      responses:
        "200":
          description: Page of records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/search:
    post:
      summary: Hybrid search
      description: |
        Vector / sparse / BM25 hybrid similarity search over a namespace,
        with offset pagination (`cursor` = zero-based offset, `limit` = page
        size; defaults to `top_k`). Wire format mirrors the SDK's
        `MemorySearchRequest`.

        **Fresh database note:** `ensure_indexes_current()` runs ONCE at server
        startup (see `src/server/bootstrap.rs`; skipped when `read_only`) and
        reconciles missing/stale text/BM25 state at that moment. It does NOT
        cover records written afterwards through the node/record API: on a
        fresh DB, `PUT /api/v2/records/…` followed by `POST /api/v2/search`
        with `text_query` fails with `text_index not found: bm25` until you
        call `POST /api/v2/maintenance/rebuild-index` (verified live-fire —
        campaign task GOV-B5). Single memory `PUT`/`DELETE` maintain the text
        index incrementally and `put_batch` rebuilds it in one pass, so this
        only bites the record-write-then-text-search flow on a fresh DB.
      operationId: searchRecords
      tags:
        - Search
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchPageRequest"
      responses:
        "200":
          description: Score-ranked page of results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchPageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/autocomplete:
    get:
      summary: IQL autocomplete
      description: Returns completion suggestions for an IQL prefix.
      operationId: iqlAutocomplete
      tags:
        - Search
      security:
        - bearerAuth: []
      parameters:
        - name: prefix
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suggestion strings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v2/audit:
    get:
      summary: Audit events
      description: Query the audit event log written when auditing is enabled.
      operationId: listAuditEvents
      tags:
        - System
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Audit log not enabled/configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/export:
    post:
      summary: Export records
      description: Export records to a JSONL file on the server, optionally filtered by namespace/metadata filter.
      operationId: exportRecords
      tags:
        - Records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: Export completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/import:
    post:
      summary: Import records
      description: 'Import records inline (`records`) or from a server-side file (`path`, JSONL default or `.vdbdump` bulk via `format: "bulk"`).'
      operationId: importRecords
      tags:
        - Records
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "200":
          description: Import completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication. If `api_key` is configured in Config,
        all requests to `/api/*` endpoints must include this header.
        Use the environment variable `VANTADB_API_KEY` to set the key.
  schemas:
...
    QueryRequest:
      description: Payload for executing an IQL query.
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: IQL query string
          example: (memory:get "agent/main" "memory-1")
...
  responses:
...
    TooManyRequests:
      description: Rate limit exceeded (configurable via rate_limit_rpm, default 100 req/min)
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitResponse"
          example:
            success: false
            error: "Too many requests. Rate limit: 100 requests per minute."
...